src/main/java/com/example/ValentinoIsgro50368/controllers/HumanoController.java
package com.example.ValentinoIsgro50368.controllers;


import com.example.ValentinoIsgro50368.entities.Humano;
import com.example.ValentinoIsgro50368.dtos.StatsDTO;
import com.example.ValentinoIsgro50368.services.HumanoService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@CrossOrigin(origins = "*")
@RequestMapping(path = "/api/adn")

public class HumanoController {

    private final HumanoService humanoService;

    public HumanoController(HumanoService humanoService) {
        this.humanoService = humanoService;
    }

    @PostMapping("/mutant")
    public ResponseEntity<?> verificarMutante(@RequestBody Humano humano) {
        try{
            String[] dnaArray = humano.getDna().toArray(new String[0]);
            if (humanoService.isMutant(dnaArray)) {
                return ResponseEntity.status(HttpStatus.OK).body("Es Mutante");
            } else {
                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(("{\"error\": \"" + "No es mutante" + "\"}"));
            }
        }
        catch (Exception e) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(("{\"error\": \"" + e.getMessage() + "\"}"));
        }
    }

    @GetMapping("/stats")
    public ResponseEntity<?> obtenerEstadisticas() {
        try{
            StatsDTO stats = humanoService.obtenerEstadisticas();
            return ResponseEntity.status(HttpStatus.OK).body(stats);
        }
        catch (Exception e) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(("{\"error\": \"" + e.getMessage() + "\"}"));
        }
    }
}

